[V1][Spec Decode] Fix eagle3 parameter definition
The redundant .type(torch.LongTensor) moves the GPU tensor to CPU
causing:

```
    targets = base + self.draft_id_to_target_id
              ~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~

RuntimeError: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu!
```

+++ vllm/model_executor/models/llama_eagle3.py
@@ -195,8 +195,7 @@ class Eagle3LlamaForCausalLM(LlamaForCausalLM):
         self.logits_processor = LogitsProcessor(self.config.draft_vocab_size,
                                                 scale=logit_scale)
         self.draft_id_to_target_id = nn.Parameter(
-            torch.zeros((self.config.draft_vocab_size),
-                        dtype=torch.long).type(torch.LongTensor),
+            torch.zeros(self.config.draft_vocab_size, dtype=torch.long),
             requires_grad=False,
         )
